refactor: BodyRequest 데이터 객체를 request 패키지 하위로 이동
