+++ src/main/java/codesquad/web/user/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package codesquad.web.user;
+package codesquad.web.user.request;
 
 public class RegisterRequest {
     private final String email;
